requirements.txt
setuptools
pyyaml
flask
flask_restful
flask_cors
argparse
requests
SQLAlchemy
psycopg2-binary
alembic
pycryptodome
validators
b2sdk